backend/requirements.txt
flask==2.2.5
flask-login==0.6.2
flask-sqlalchemy==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==2.2.3
openai==1.12.0
python-jose==3.3.0 